requirements.txt
blinker==1.6.2
boto3==1.26.148
botocore==1.29.148
click==8.1.3
Flask==2.3.2
Flask-DotEnv==0.1.2
Flask-SQLAlchemy==3.0.3
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
psycopg2-binary==2.9.6
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.6.3
urllib3==1.26.16
uuid==1.30
Werkzeug==2.3.4
